[MOV] chart: rename `title` files to `chart_title`
Task: 4575651
