+++ src/components/side_panel/chart/building_blocks/chart_title/chart_title.xml
